pyproject.toml
[tool.poetry]
name = "python-drf"
version = "0.1.0"
description = ""
authors = ["Sasha <sasaserbinena@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.16.0"
mysqlclient = "^2.2.7"
poetry-core = "^2.1.3"
dotenv = "^0.9.9"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.1"
pillow = "^11.3.0"


[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
